setup to user add on cloudinary

+++ src/utils/cloudinary.js
@@ -1,26 +1,35 @@
-import { v2 as cloudinary } from "cloudinary"
-import fs from "fs"
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_SECRET
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
-    if (!localFilePath) return null
-    // upload the file on cloudinary
-    const response = await cloudinary.uploader.upload(localFilePath, {
-      resource_type: "auto",
-    })
-    //file has been uploaded successfully
-    console.log("file is uploaded on cloudinary ", response.url);
-    return response;
-  } catch (error) {
-     fs.unlinkSync(localFilePath)
-     return null
-  }
-}
+    if (!localFilePath) return null;
 
-export {uploadOnCloudinary}
+    const result = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
+
+    console.log("✅ File uploaded on Cloudinary:", result.url);
+
+    console.log("Uploading from path:", localFilePath);
+    // delete local file after successful upload
+    fs.unlinkSync(localFilePath);
+    return result;
+  } catch (error) {
+    console.error("❌ Cloudinary upload error:", error.message);
+
+    // delete local file only if it exists
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
